requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
dlib==19.19.0
dnspython==1.16.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
Flask-Cors==3.0.8
gevent==1.4.0
greenlet==0.4.15
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
numpy==1.18.2
opencv-python==4.2.0.32
Pillow==7.0.0
playsound==1.2.2
protobuf==3.6.1
psycopg2-binary==2.8.4
python-dotenv==0.13.0
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
soupsieve==2.0
tqdm==4.46.1
urllib3==1.25.8
vext==0.7.3
Werkzeug==1.0.1
